delete signup component and make it in page
