--- client/src/Components/Sign Up/style.css
@@ -1,32 +1,0 @@
-.input-conatiner{
-display: flex;
-flex-direction: column;
-width: 80%;
-
-
-margin-top: 2%;
-padding-top: 50px;
-padding: 50px;
-
-text-align: center;
-align-items: center;
-
-}
-.container-main{
-    display: flex;
-    flex-direction: row;
-    width: 100%;
-    height: 100vh;
-    
-}
-.image{
-    width: 100%;
-}
-.signupImage{
-    clip-path: polygon(0 0, 100% 0, 69% 100%, 0% 100%);
-    height: 100vh;
-    width: 100%;
-    background-size: cover;
-    object-fit: cover;
-    
-}
